feat: refactor implementation for conditional compilation
